requirements.txt
requests
django==2.1.*
gunicorn
djangorestframework
